Support for 16 OCMBs
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/80364
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/occ_405/amec/amec_health.h
@@ -5,7 +5,7 @@
 /*                                                                        */
 /* OpenPOWER OnChipController Project                                     */
 /*                                                                        */
-/* Contributors Listed Below - COPYRIGHT 2011,2017                        */
+/* Contributors Listed Below - COPYRIGHT 2011,2019                        */
 /* [+] International Business Machines Corp.                              */
 /*                                                                        */
 /*                                                                        */
@@ -49,7 +49,7 @@ void amec_health_check_dimm_temp(void);
 void amec_health_check_dimm_timeout(void);
 void amec_mem_mark_logged(uint8_t i_cent,
                           uint8_t i_dimm,
-                          uint8_t* i_clog_bitmap,
+                          uint16_t* i_clog_bitmap,
                           uint8_t* i_dlog_bitmap);
 void amec_health_check_vrm_vdd_temp(const sensor_t *i_sensor);
 void amec_health_check_vrm_vdd_temp_timeout(void);
